🚚 Moved tests to tests folder
