+++ package.json
@@ -4,7 +4,7 @@
   "description": "Prompt for a user response using reactions or a massage.",
   "main": "./src/index.js",
   "scripts": {
-    "test": "jest"
+    "test": "jest --verbose"
   },
   "keywords": [
     "discord",
@@ -33,5 +33,8 @@
     "eslint": "^6.0.1",
     "jest": "^24.8.0",
     "typescript": "^3.5.3"
+  },
+  "jest": {
+    "testMatch": [ "**/tests/**/*.[jt]s?(x)", "**/?(*.)+(spec|test).[jt]s?(x)" ]
   }
 }
